test_api/requirements.txt
anyio==3.7.0
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
fastapi==0.99.1
h11==0.14.0
idna==3.4
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.10
PyJWT==2.7.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0

